feat(src): rename `cloud-instance-metadata` into `cloud-metadata` in codebase
