+++ src/__tests__/unit/lib/cloud-metadata/index.test.ts
@@ -4,7 +4,7 @@ import {ERRORS} from '../../../../util/errors';
 
 const {InputValidationError, UnsupportedValueError} = ERRORS;
 
-describe('lib/cloud-instance-metadata:', () => {
+describe('lib/cloud-metadata:', () => {
   describe('CloudMetadata', () => {
     const cloudMetadata = CloudMetadata();
