+++ src/lib/cloud-metadata/aws-instances.csv
