+++ src/lib/cloud-metadata/azure-instances.csv
