+++ src/lib/cloud-metadata/config.ts
