+++ src/lib/cloud-metadata/index.ts
