+++ src/lib/cloud-metadata/types.ts
