+++ src/lib/index.ts
@@ -1,4 +1,4 @@
-export {CloudMetadata} from './cloud-instance-metadata';
+export {CloudMetadata} from './cloud-metadata';
 export {EMem} from './e-mem';
 export {ENet} from './e-net';
 export {Sci} from './sci';
