+++ src/lib/sci-m/README.md
@@ -18,7 +18,7 @@ Not Needed
 - `resources-total`: the total number of resources available
 - `duration`: the amount of time covered by an observation, in this context it is used as the share of the total life span of the hardware reserved for use by an application, in seconds.
 
-> Note that if you have a plugin pipeline that adds `vcpus-allocated` and `vcpus-total` to each observation, such as the `cloud-instance-metadata` plugin, those values will be used **in preference** to the given `resources-reserved` and `resources-total` fields.
+> Note that if you have a plugin pipeline that adds `vcpus-allocated` and `vcpus-total` to each observation, such as the `cloud-metadata` plugin, those values will be used **in preference** to the given `resources-reserved` and `resources-total` fields.
 
 ## Returns
